Move test parsers to tests.py
